refactor(launch/manager): move tests out of input/mod.rs into submodules
The `input/mod.rs` `#[cfg(test)] mod tests` block held ~1800 lines
of test cases covering every stage. That concentration forced
future maintainers to navigate a large mixed-responsibility file
and surfaced in static analysis as a cluster of duplicate-binding
warnings localized to one place.

Disperse the tests so each submodule owns the tests for the
behavior it owns:

- list-stage tests -> input/list.rs
- editor-stage tests -> input/editor.rs
- save-flow tests -> input/save.rs
- create-wizard tests -> input/prelude.rs
- mouse tests stay in input/mouse.rs

input/mod.rs retains only the genuine cross-flow test that drives
multiple stages in one case (create-mode rename + save). Shared
helpers `key()` and `mount()` are lifted to a `test_support`
sibling module since they show up in 3+ submodules; submodule-
local helpers (setup_with_workspace, press_s, etc.) stay co-
located with their tests.

No behavior change. Test count unchanged (829 -> 829).

Addresses finding #6 of the second-pass PR #166 review.
