+++ src/launch/manager/input/list.rs
@@ -192,3 +192,423 @@ pub(super) fn handle_list_modal(state: &mut ManagerState<'_>, key: KeyEvent) {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    //! List-stage tests: row-0 (current dir) gating, Enter routing,
+    //! `o`-key resolver to GitHub URLs, and the `GithubPicker` modal.
+    use super::super::super::state::{ManagerStage, ManagerState, Modal, ToastKind};
+    use super::super::test_support::{key, mount};
+    use super::InputOutcome;
+    use crate::config::AppConfig;
+    use crate::launch::manager::input::handle_key;
+    use crate::paths::JackinPaths;
+    use crate::workspace::WorkspaceConfig;
+    use crossterm::event::KeyCode;
+    use tempfile::TempDir;
+
+    /// Build a git repo under `root` with a `github.com` origin remote on
+    /// `branch`. Returns the path so callers can use it as a mount src.
+    fn make_github_repo(root: &std::path::Path, name: &str, branch: &str) -> std::path::PathBuf {
+        let path = root.join(name);
+        let git_dir = path.join(".git");
+        std::fs::create_dir_all(&git_dir).unwrap();
+        std::fs::write(git_dir.join("HEAD"), format!("ref: refs/heads/{branch}\n")).unwrap();
+        std::fs::write(
+            git_dir.join("config"),
+            format!("[remote \"origin\"]\n    url = git@github.com:owner/{name}.git\n"),
+        )
+        .unwrap();
+        path
+    }
+
+    /// Helper: seed an AppConfig + ManagerState with `ws` as a saved workspace,
+    /// cwd far away so selection lands on row 1 (the saved workspace).
+    fn list_state_selecting_ws(
+        ws: WorkspaceConfig,
+    ) -> (ManagerState<'static>, AppConfig, JackinPaths, TempDir) {
+        let tmp = tempfile::tempdir().unwrap();
+        let paths = JackinPaths::for_tests(tmp.path());
+        paths.ensure_base_dirs().unwrap();
+        let mut config = AppConfig::default();
+        config.workspaces.insert("demo".into(), ws);
+        let mut state = ManagerState::from_config(&config, tmp.path());
+        state.selected = 1; // force selection onto the saved workspace row
+        (state, config, paths, tmp)
+    }
+
+    /// Current-directory row (index 0) must reject the `e` edit shortcut and
+    /// the `d` delete shortcut with a toast, without entering the Editor or
+    /// ConfirmDelete stages. Paired with the render-side assertion that row 0
+    /// is labelled "Current directory".
+    #[test]
+    fn current_directory_row_rejects_edit_and_delete() {
+        let tmp = tempfile::tempdir().unwrap();
+        let paths = JackinPaths::for_tests(tmp.path());
+        paths.ensure_base_dirs().unwrap();
+        let cwd = tmp.path();
+
+        // Minimal config with one saved workspace so the list has a non-
+        // trivial shape (current-dir + one saved + sentinel).
+        let mut config = AppConfig::default();
+        config.workspaces.insert(
+            "some-ws".into(),
+            WorkspaceConfig {
+                workdir: "/unrelated".into(),
+                mounts: vec![],
+                allowed_agents: vec![],
+                default_agent: None,
+                last_agent: None,
+                env: std::collections::BTreeMap::new(),
+                agents: std::collections::BTreeMap::new(),
+            },
+        );
+        let mut state = ManagerState::from_config(&config, cwd);
+        // cwd is unrelated to /unrelated, so preselect falls back to row 0.
+        assert_eq!(state.selected, 0);
+
+        // Press `e` — must produce a toast and remain in the List stage.
+        handle_key(
+            &mut state,
+            &mut config,
+            &paths,
+            cwd,
+            key(KeyCode::Char('e')),
+        )
+        .unwrap();
+        assert!(
+            matches!(&state.stage, ManagerStage::List),
+            "e on row 0 must not open the Editor; got {:?}",
+            state.stage
+        );
+        let toast = state.toast.as_ref().expect("edit rejection must toast");
+        assert!(
+            matches!(toast.kind, ToastKind::Error),
+            "edit rejection must be an error toast"
+        );
+        assert!(
+            toast.message.contains("edit"),
+            "toast should mention edit: {}",
+            toast.message
+        );
+        state.toast = None;
+
+        // Press `d` — must produce a toast and remain in the List stage
+        // (no ConfirmDelete transition).
+        handle_key(
+            &mut state,
+            &mut config,
+            &paths,
+            cwd,
+            key(KeyCode::Char('d')),
+        )
+        .unwrap();
+        assert!(
+            matches!(&state.stage, ManagerStage::List),
+            "d on row 0 must not open ConfirmDelete; got {:?}",
+            state.stage
+        );
+        let toast = state.toast.as_ref().expect("delete rejection must toast");
+        assert!(
+            matches!(toast.kind, ToastKind::Error),
+            "delete rejection must be an error toast"
+        );
+        assert!(
+            toast.message.contains("delete"),
+            "toast should mention delete: {}",
+            toast.message
+        );
+    }
+
+    /// Enter on row 0 returns `LaunchCurrentDir`; Enter on row 1 returns
+    /// `LaunchNamed(<name>)`. Pins the index arithmetic that maps list-row
+    /// indices to launch targets.
+    #[test]
+    fn enter_on_current_directory_returns_launch_current_dir() {
+        let tmp = tempfile::tempdir().unwrap();
+        let paths = JackinPaths::for_tests(tmp.path());
+        paths.ensure_base_dirs().unwrap();
+        let cwd = tmp.path();
+
+        let mut config = AppConfig::default();
+        config.workspaces.insert(
+            "alpha".into(),
+            WorkspaceConfig {
+                workdir: "/alpha".into(),
+                mounts: vec![],
+                allowed_agents: vec![],
+                default_agent: None,
+                last_agent: None,
+                env: std::collections::BTreeMap::new(),
+                agents: std::collections::BTreeMap::new(),
+            },
+        );
+        let mut state = ManagerState::from_config(&config, cwd);
+        state.selected = 0;
+        let outcome =
+            handle_key(&mut state, &mut config, &paths, cwd, key(KeyCode::Enter)).unwrap();
+        assert!(
+            matches!(outcome, InputOutcome::LaunchCurrentDir),
+            "row 0 Enter must produce LaunchCurrentDir"
+        );
+
+        state.selected = 1;
+        let outcome =
+            handle_key(&mut state, &mut config, &paths, cwd, key(KeyCode::Enter)).unwrap();
+        match outcome {
+            InputOutcome::LaunchNamed(name) => assert_eq!(name, "alpha"),
+            other => panic!("row 1 Enter must produce LaunchNamed(\"alpha\"); got {other:?}"),
+        }
+    }
+
+    // ── List-view `o` key → GitHub resolver + picker ──────────────────
+
+    #[test]
+    fn resolve_github_mounts_returns_one_per_github_repo() {
+        // A workspace with two github mounts + one folder + one gitlab repo
+        // should yield exactly two picker choices.
+        let tmp = tempfile::tempdir().unwrap();
+        let repo_a = make_github_repo(tmp.path(), "repo-a", "main");
+        let repo_b = make_github_repo(tmp.path(), "repo-b", "dev");
+        let plain = tmp.path().join("plain");
+        std::fs::create_dir(&plain).unwrap();
+        // Gitlab repo should be skipped.
+        let gitlab = tmp.path().join("gl");
+        let gl_git = gitlab.join(".git");
+        std::fs::create_dir_all(&gl_git).unwrap();
+        std::fs::write(gl_git.join("HEAD"), "ref: refs/heads/main\n").unwrap();
+        std::fs::write(
+            gl_git.join("config"),
+            "[remote \"origin\"]\n    url = git@gitlab.com:owner/repo.git\n",
+        )
+        .unwrap();
+
+        let ws = WorkspaceConfig {
+            workdir: String::new(),
+            mounts: vec![
+                mount(repo_a.to_str().unwrap(), "/a"),
+                mount(plain.to_str().unwrap(), "/p"),
+                mount(repo_b.to_str().unwrap(), "/b"),
+                mount(gitlab.to_str().unwrap(), "/g"),
+            ],
+            allowed_agents: vec![],
+            default_agent: None,
+            last_agent: None,
+            env: std::collections::BTreeMap::new(),
+            agents: std::collections::BTreeMap::new(),
+        };
+
+        let choices = crate::launch::manager::github_mounts::resolve_for_workspace(&ws);
+        assert_eq!(choices.len(), 2);
+        // URLs track the HEAD ref per-repo.
+        let urls: Vec<&str> = choices.iter().map(|c| c.url.as_str()).collect();
+        assert!(urls.contains(&"https://github.com/owner/repo-a/tree/main"));
+        assert!(urls.contains(&"https://github.com/owner/repo-b/tree/dev"));
+        // Branch label matches Named variant.
+        let branches: Vec<&str> = choices.iter().map(|c| c.branch.as_str()).collect();
+        assert!(branches.contains(&"main"));
+        assert!(branches.contains(&"dev"));
+    }
+
+    #[test]
+    fn list_o_with_single_github_mount_has_one_resolved_url() {
+        // Resolver-side check — we can't cleanly assert `open::that_detached`
+        // ran, but we can pin that there's exactly one URL to hand to it so
+        // the 1-mount branch's immediate-open path is taken.
+        let tmp = tempfile::tempdir().unwrap();
+        let repo = make_github_repo(tmp.path(), "solo", "trunk");
+        let ws = WorkspaceConfig {
+            workdir: String::new(),
+            mounts: vec![mount(repo.to_str().unwrap(), "/solo")],
+            allowed_agents: vec![],
+            default_agent: None,
+            last_agent: None,
+            env: std::collections::BTreeMap::new(),
+            agents: std::collections::BTreeMap::new(),
+        };
+        let choices = crate::launch::manager::github_mounts::resolve_for_workspace(&ws);
+        assert_eq!(choices.len(), 1);
+        assert_eq!(choices[0].url, "https://github.com/owner/solo/tree/trunk");
+    }
+
+    #[test]
+    fn list_o_with_multiple_github_mounts_opens_picker() {
+        let tmp = tempfile::tempdir().unwrap();
+        let repo_a = make_github_repo(tmp.path(), "repo-a", "main");
+        let repo_b = make_github_repo(tmp.path(), "repo-b", "main");
+        let ws = WorkspaceConfig {
+            workdir: String::new(),
+            mounts: vec![
+                mount(repo_a.to_str().unwrap(), "/a"),
+                mount(repo_b.to_str().unwrap(), "/b"),
+            ],
+            allowed_agents: vec![],
+            default_agent: None,
+            last_agent: None,
+            env: std::collections::BTreeMap::new(),
+            agents: std::collections::BTreeMap::new(),
+        };
+        let (mut state, mut config, paths, tmp) = list_state_selecting_ws(ws);
+
+        handle_key(
+            &mut state,
+            &mut config,
+            &paths,
+            tmp.path(),
+            key(KeyCode::Char('o')),
+        )
+        .unwrap();
+
+        match &state.list_modal {
+            Some(Modal::GithubPicker { state: picker }) => {
+                assert_eq!(picker.choices.len(), 2);
+            }
+            other => panic!("expected GithubPicker modal; got {other:?}"),
+        }
+    }
+
+    #[test]
+    fn list_o_with_zero_github_mounts_shows_toast() {
+        let tmp_src = tempfile::tempdir().unwrap();
+        let plain = tmp_src.path().join("plain");
+        std::fs::create_dir(&plain).unwrap();
+        let ws = WorkspaceConfig {
+            workdir: String::new(),
+            mounts: vec![mount(plain.to_str().unwrap(), "/p")],
+            allowed_agents: vec![],
+            default_agent: None,
+            last_agent: None,
+            env: std::collections::BTreeMap::new(),
+            agents: std::collections::BTreeMap::new(),
+        };
+        let (mut state, mut config, paths, tmp) = list_state_selecting_ws(ws);
+
+        handle_key(
+            &mut state,
+            &mut config,
+            &paths,
+            tmp.path(),
+            key(KeyCode::Char('o')),
+        )
+        .unwrap();
+
+        assert!(
+            state.list_modal.is_none(),
+            "no modal should open when there are no github mounts"
+        );
+        let toast = state.toast.as_ref().expect("expected a toast");
+        assert!(
+            toast.message.contains("no GitHub URL"),
+            "toast should explain the no-mounts state: {}",
+            toast.message
+        );
+    }
+
+    #[test]
+    fn list_o_on_row_zero_toasts_no_workspace_selected() {
+        // Row 0 is the synthetic "Current directory" — no saved workspace
+        // to read mounts from; hint should nudge the operator, not crash.
+        let tmp = tempfile::tempdir().unwrap();
+        let paths = JackinPaths::for_tests(tmp.path());
+        paths.ensure_base_dirs().unwrap();
+        let mut config = AppConfig::default();
+        config.workspaces.insert(
+            "demo".into(),
+            WorkspaceConfig {
+                workdir: String::new(),
+                mounts: vec![],
+                allowed_agents: vec![],
+                default_agent: None,
+                last_agent: None,
+                env: std::collections::BTreeMap::new(),
+                agents: std::collections::BTreeMap::new(),
+            },
+        );
+        let mut state = ManagerState::from_config(&config, tmp.path());
+        state.selected = 0;
+
+        handle_key(
+            &mut state,
+            &mut config,
+            &paths,
+            tmp.path(),
+            key(KeyCode::Char('o')),
+        )
+        .unwrap();
+
+        let toast = state.toast.as_ref().expect("expected a toast");
+        assert!(toast.message.contains("no workspace selected"));
+        assert!(state.list_modal.is_none());
+    }
+
+    #[test]
+    fn picker_commit_closes_list_modal_and_clears_state() {
+        // Seed the state directly with an open GithubPicker, then commit.
+        // We can't assert `open::that_detached` ran, but we *can* pin that
+        // the modal closes (no lingering state) and no error toast appears
+        // when the underlying call path doesn't error out synchronously.
+        use crate::launch::widgets::github_picker::{GithubChoice, GithubPickerState};
+        let tmp = tempfile::tempdir().unwrap();
+        let paths = JackinPaths::for_tests(tmp.path());
+        paths.ensure_base_dirs().unwrap();
+        let mut config = AppConfig::default();
+        let mut state = ManagerState::from_config(&config, tmp.path());
+        // Use an unreachable file:// URL so `open::that_detached` is a
+        // cheap no-op on most platforms (still spawns the browser handler
+        // but doesn't block on network).
+        state.list_modal = Some(Modal::GithubPicker {
+            state: GithubPickerState::new(vec![GithubChoice {
+                src: "/tmp/a".into(),
+                branch: "main".into(),
+                url: "file:///dev/null".into(),
+            }]),
+        });
+
+        handle_key(
+            &mut state,
+            &mut config,
+            &paths,
+            tmp.path(),
+            key(KeyCode::Enter),
+        )
+        .unwrap();
+
+        assert!(
+            state.list_modal.is_none(),
+            "picker Enter must close the modal"
+        );
+    }
+
+    #[test]
+    fn picker_esc_closes_without_opening_url() {
+        use crate::launch::widgets::github_picker::{GithubChoice, GithubPickerState};
+        let tmp = tempfile::tempdir().unwrap();
+        let paths = JackinPaths::for_tests(tmp.path());
+        paths.ensure_base_dirs().unwrap();
+        let mut config = AppConfig::default();
+        let mut state = ManagerState::from_config(&config, tmp.path());
+        state.list_modal = Some(Modal::GithubPicker {
+            state: GithubPickerState::new(vec![GithubChoice {
+                src: "/tmp/a".into(),
+                branch: "main".into(),
+                url: "https://github.com/owner/repo/tree/main".into(),
+            }]),
+        });
+
+        handle_key(
+            &mut state,
+            &mut config,
+            &paths,
+            tmp.path(),
+            key(KeyCode::Esc),
+        )
+        .unwrap();
+
+        assert!(state.list_modal.is_none());
+        assert!(
+            state.toast.is_none(),
+            "Esc must not toast: {:?}",
+            state.toast
+        );
+    }
+}
